backend/requirements.txt
# ============================================================================
# ARKEYEZDOC v2.0 - Requirements (WebSocket + ERPNext Integration)
# ============================================================================

# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0  # NEW: WebSocket support

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0  # NEW: Environment variables

# Machine Learning
tensorflow==2.19.0
numpy==1.26.4
pillow==10.1.0

# OCR & NLP
easyocr==1.7.0
scikit-learn==1.3.2

# PDF Processing
PyMuPDF==1.23.8

# HTTP Requests (for ERPNext)
requests==2.31.0

# Database
# sqlite3 is built-in

# Utilities
pydantic==2.5.0

# Logging
loguru==0.7.2
torch==2.2.0
torchvision==0.17.0